refactor: rename interface files and exported function in cityValidator, add JSDoc typings to ISubscriptionRepository.js
