+++ src/api/domain/providers/IEmailProvider.js
@@ -1,7 +1,7 @@
-class EmailProvider {
+class IEmailProvider {
   async sendEmail({ _to, _subject, _html }) {
     throw new Error('sendEmail() must be implemented');
   }
 }
 
-module.exports = EmailProvider;
+module.exports = IEmailProvider;
