+++ src/api/infrastructure/compositionRoot/validators/validateSubscriptionInput.js
@@ -1,11 +1,11 @@
 const {
   createValidateSubscriptionInput,
 } = require('../../../presentation/middlewares/createValidateSubscriptionInput');
-const { createValidator } = require('./cityValidator');
+const { createCityValidator } = require('./cityValidator');
 const { isValidEmail } = require('../../validation/validateSubscriptionFields');
 
 const validateSubscriptionInput = createValidateSubscriptionInput(
-  createValidator(),
+  createCityValidator(),
   isValidEmail,
 );
