+++ src/api/infrastructure/providers/WeatherAPIProvider.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const AbstractWeatherProvider = require('../../domain/providers/AbstractWeatherProvider');
+const AbstractWeatherProvider = require('../../domain/providers/IWeatherProvider');
 const logProviderResponse = require('../logging/logProviderResponse');
 
 class WeatherAPIProvider extends AbstractWeatherProvider {
